PrintClient server: bring your own user
