+++ src/Mapbender/CoreBundle/Component/Element.php
@@ -13,6 +13,8 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
  * Base class for all Mapbender elements.
@@ -381,6 +383,24 @@ abstract class Element
     public function httpAction($action)
     {
         throw new NotFoundHttpException('This element has no Ajax handler.');
+    }
+
+    /**
+     * Gets the user making the current request. Useful for httpAction.
+     * Return type is a bit flexible, @see AbstractToken::setUser
+     *
+     * @return UserInterface|string|object|null
+     */
+    protected function getUser()
+    {
+        /** @var TokenStorageInterface $tokenStorage */
+        $tokenStorage = $this->container->get('security.token_storage');
+        $token = $tokenStorage->getToken();
+        if ($token) {
+            return $token->getUser();
+        } else {
+            return null;
+        }
     }
 
     /**
